Added S4U client side
Currently just handles S4U to self within the client
Fixed all the enums because the encoding is a little looser than AD prefers

+++ Kerberos.NET/Asn1/BitOperation.cs
@@ -15,7 +15,7 @@ namespace Kerberos.NET
         {
             var longVal = (object)val;
 
-            return AsReadOnly((long)longVal, littleEndian: littleEndian);
+            return AsReadOnly((int)longVal, littleEndian: littleEndian);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
